+++ src/sdc11073/xmlparsing.py
@@ -71,3 +71,22 @@ def copy_node(node: etree_._Element, method=copy.deepcopy) -> etree_._Element:
     new_report = method(root_tree.getroot())
     ns_map = {k: v for k, v in node.nsmap.items() if k}  # filter for default namespace
     return new_report.xpath(root_tree.getpath(node), namespaces=ns_map)[0]
+
+
+def compare_xml_extension(left: etree_._Element, right: etree_._Element) -> bool:
+    if left.tag != right.tag:  # compare expanded names
+        return False
+    if dict(left.attrib) != dict(right.attrib):  # unclear how lxml _Attrib compares
+        return False
+
+    # ignore comments
+    left_children = [child for child in left if not isinstance(child, etree_._Comment)]
+    right_children = [child for child in right if not isinstance(child, etree_._Comment)]
+
+    if len(left_children) != len(right_children):  # compare children count
+        return False
+    if len(left_children) == 0 and len(right_children) == 0:
+        if left.text != right.text:  # mixed content is not allowed. only compare text if there are no children
+            return False
+
+    return all(map(compare_xml_extension, left_children, right_children))  # compare children but keep order