+++ tests/test_pmtypes.py
@@ -1,5 +1,7 @@
 import unittest
 
+import lxml
+
 from sdc11073 import pmtypes
 
 
@@ -20,7 +22,7 @@ class TestPmtypes(unittest.TestCase):
         c2.Translation.append(pmtypes.T_Translation(41))
         self.assertEqual(c2, 41)
         c3 = pmtypes.CodedValue(42)
-        c3.Translation.append(pmtypes.T_Translation(41)) # same translation as c2
+        c3.Translation.append(pmtypes.T_Translation(41))  # same translation as c2
         self.assertEqual(c2, c3)
 
     def test_base_demographics(self):
@@ -31,3 +33,35 @@ class TestPmtypes(unittest.TestCase):
         self.assertEqual(bd.Middlename, ['foo'])
         bd = pmtypes.BaseDemographics(middlenames=['foo', 'bar'])
         self.assertEqual(bd.Middlename, ['foo', 'bar'])
+
+
+class TestExtensions(unittest.TestCase):
+
+    def test_compare_extensions(self):
+        xml = b"""
+        <pm:Identification xmlns:pm="__BICEPS_ParticipantModel__"
+                           Root="urn:uuid:90beab82-f160-4e2f-b3b2-ed8cfcf5e205"
+                           Extension="123.234.424">
+            <ext:Extension xmlns:ext="__ExtensionPoint__">
+                <foo someattr="somevalue"/>
+                    <foo_child childattr="somechild"/>
+                <bar anotherattr="differentvalue"/>
+            </ext:Extension>
+        </pm:Identification>
+        """
+        inst1 = pmtypes.InstanceIdentifier.fromNode(lxml.etree.fromstring(xml))
+        inst2 = pmtypes.InstanceIdentifier.fromNode(lxml.etree.fromstring(xml))
+        self.assertEqual(inst1, inst2)
+
+        another_xml = b"""
+                <pm:Identification xmlns:pm="__BICEPS_ParticipantModel__"
+                                   Root="urn:uuid:90beab82-f160-4e2f-b3b2-ed8cfcf5e205"
+                                   Extension="123.234.424">
+                    <ext:Extension xmlns:ext="__ExtensionPoint__">
+                        <foo someattr="somevalue"/>
+                        <bar anotherattr="differentvalue2"/>
+                    </ext:Extension>
+                </pm:Identification>
+                """
+        inst2 = pmtypes.InstanceIdentifier.fromNode(lxml.etree.fromstring(another_xml))
+        self.assertNotEqual(inst1, inst2)